预设部门的编辑与删除禁止

+++ src/components/System/Account/Department/DepartmentMain/DepartmentTable.jsx
@@ -75,14 +75,15 @@ class DepartmentTable extends Component {
       selectedDepartment:selectedRows
     })
   }
-  onWarningTipOK = () => {//删除部门，todo
+  onWarningTipOK = () => {//删除部门，拒绝编辑部门提示框
     const { selectedDepartment,deleteDepartment } = this.props;
-    const selectedDepartmentHasChild = selectedDepartment.map(e=>e.hasChildren).some(e=>!!e);
+    const selectedDepartmentHasChild = selectedDepartment.map(e=>e.hasChildren).some(e=>!!e);//有子部门不得删除
+    const forbiddenDelete = selectedDepartment.some(e=>e.departmentSource === 0); //预设电站不得删除
     this.setState({
       showWarningTip:false,
       hiddenWarningTipCancelText: false
     })
-    selectedDepartmentHasChild || deleteDepartment({
+    forbiddenDelete || selectedDepartmentHasChild || deleteDepartment({
       departmentId: selectedDepartment.map(e=>e.departmentId).join(','),
       enterpriseId: this.props.enterpriseId
     });
@@ -124,14 +125,30 @@ class DepartmentTable extends Component {
   departmentHandle = (value) => {//编辑，删除，分配用户/电站
     const { selectedDepartment } = this.props;
     if(value==='edit'){
-      this.props.changeDepartmentStore({
-        showPage: 'edit',
-        departmentDetail: selectedDepartment[0],
-      });
+      const forbiddenEdit = selectedDepartment.some(e=>e.departmentSource === 0);
+      if(forbiddenEdit){
+        this.setState({
+          showWarningTip: true,
+          warningTipText: '不得编辑预设部门!',
+          hiddenWarningTipCancelText: true
+        })
+      }else{
+        this.props.changeDepartmentStore({
+          showPage: 'edit',
+          departmentDetail: selectedDepartment[0],
+        });
+      }
     }else if(value==='delete'){
       const selectedDepartmentHasChild = selectedDepartment.map(e=>e.hasChildren).some(e=>!!e);
       const selectedDepartmentHasMember = selectedDepartment.map(e=>e.hasMember).some(e=>!!e);
-      if(selectedDepartmentHasChild){
+      const forbiddenDelete = selectedDepartment.some(e=>e.departmentSource === 0)
+      if(forbiddenDelete){
+        this.setState({
+          showWarningTip: true,
+          warningTipText: '不得删除预设部门!',
+          hiddenWarningTipCancelText: true
+        })
+      }else if(selectedDepartmentHasChild){
         this.setState({
           showWarningTip: true,
           warningTipText: '请先删除子部门!',
@@ -169,7 +186,7 @@ class DepartmentTable extends Component {
       editable = selectedDepartment.length === 1;
       [deletable, userAssignable, staionAssignable] = [true,true,true];
     }       
-    return (<Select onChange={this.departmentHandle} placeholder={'操作'} dropdownMatchSelectWidth={false} dropdownClassName={styles.handleDropdown}>
+    return (<Select onChange={this.departmentHandle} placeholder="操作" value="操作" dropdownMatchSelectWidth={false} dropdownClassName={styles.handleDropdown}>
       <Option value="edit" disabled={!editable} >编辑</Option>
       <Option value="delete" disabled={!deletable} >删除</Option>
       <Option value="assignUser" disabled={!userAssignable} >分配用户</Option>
